backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.2.1
aiohttp==3.9.1
Pillow==10.1.0
numpy==1.24